requirements.txt
numpy==1.19.5
scipy==1.5.4
scikit-learn==0.23.2
joblib==0.17.0
tensorflow-gpu==2.4.1
matplotlib~=3.3.4
wfdb~=3.3.0
h5py~=2.10.0